services/analytics/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2
httpx==0.25.2
requests==2.31.0
tenacity==8.2.3
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.18.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
opentelemetry-instrumentation-redis==0.42b0

prometheus-client==0.19.0
